requirements.txt
mysql-connector-python==9.4.0
passlib==1.7.4
